split activate function into its own file
